Fix puppet-lint violations to improve quality score
Increasing the quality score should hopefully encourage more people
to feel comfortable adopting the module - and the more people using
it, the healthier the module!

The easy way to fix the violations: http://puppet-lint.com/
I just add that to run as a pre-commit hook and it keeps me honest.