Flip default so newer releases will use latest packages
Having the default be used for the older releases (5.0 and 6.0) meant
that in the future we would have to continue to update to keep up.
By flipping the default, the only time we need special processing
is for older releases which likely won't be changing anyway.

I also changed the service name to be dependent on the runner release
we are using.